Try snapping to grid center
If we are using grid spaces, then if the token moves diagonally in 2-D, the 3-D measure may be inconsistent. This appears to be because the projected hypotenuse endpoint is no longer centered in the grid, causing rounding to push the distance calculation off depending on cardinal direction of the move.