🐛 fix|Teleport|Switch to holding F down to trigger teleport
Hitting the teleport key was not ideal b/c it interrupted the drag in weird ways, causing the drag to still continue after the teleport. Instead, use a modifier key when trigging the token move.

Don't use arrow key because it screws up use of arrow keys on the canvas.

Wrap _onMoveKeyDown to handle the event from the drag left drop.