💡 refactor|Waypoints|Add new elevation props to waypoints
Override _addWaypoint so the waypoint properties can be added prior to measurement.

Add _forceToGround to track if this ruler segment should start at the ground level.

Add _userElevationIncrements to track user increases from the start of this segment.

Add _terrainElevation to track ground elevation at the waypoint.

Add _prevElevation to save the elevation from the previous waypoint to this one.