🐛 fix|MovePenalty|PF2e not displaying speed correctly
MovePenalty is handling the speed calc, so use the SPEED functions to access token speed. This fixes the issue with pf2e speeds. But it is not possible to easily reset the token speed in pf2e, so penalty calcs for regions will likely be off. Needs more work to handle that.